Keep crust-render a single source file: fold the Ptex decoder into main.rs
`crust-render` is documented as the thin CLI binary whose `main.rs` is the only
file, and the Ptex work broke that by adding `mod ptex_color;` and a second
source file. Moved the decoder into `main.rs`, next to the EXR and LDR image
decoders — everything that knows a file format already lives there, so it is
where a reader would look for it anyway.

Considered giving the decoder its own crate instead, which would have kept
`main.rs` shorter. Not done: the invariant is about this crate staying a single
file, a sixth workspace crate for ~200 lines of format adapter is a heavier
structural change than the problem warrants, and the `AssetLoader` seam is
explicitly documented as the CLI's side of the split.

One real hazard surfaced in the move rather than being carried over silently:
`main.rs` glob-imports `exr::prelude`, whose one-parameter `Result<T>` alias
shadows `std::result::Result`, so `PtexColor::open`'s signature no longer
compiled as written. Spelled out and commented at the definition.

Also pins the `ptex-rs` git dependency by `rev`. The crate is unpublished and
`Cargo.lock` is gitignored, so a fresh clone was taking whatever `main` pointed
at — reproducible only by luck.

Pure code motion, and checked as such: the isLavaRocks Ptex render is
bit-identical across the move (0 of 309 400 pixels differ). 234 tests pass, and
clippy reports nothing new (`crust-render/src/main.rs` keeps its one pre-existing
redundant-closure warning in a test, unchanged in count).
